Change `GraphIR` optimisation approach + remove `TrainerBuilder` (#441)
Previously would fuse
```
Concat(SparseAffineActivate(weights, biases, input1), SparseAffineActivate(weights, biases, input2))
-> SparseAffineDualActivate(weights, biases, input1, input2)
```
However this is inflexible (each `SparseAffineActivate` must have the same weights and biases), and trying to make it more flexible causes the complexity of the `SparseAffineDualActivate` operation and optimistion passes on it to explode, in addition to requiring the underlying instructions for `SparseAffineActivate` to be strided.

We are aware of what NNUE architectures look like, and we know that batch sizes are generally very large, so that operations on non-batched nodes are almost always negligible.

Then we now perform the following optimisations instead
```
Concat(PairwiseMul(a), PairwiseMul(b)) -> FusedPairwiseMulConcat(a, b) if a != b
```
Or, if there are no `PairwiseMul`s
```
Matmul(unbatched, Concat(batched_a, batched_b))
-> Add(Matmul(Slice(unbatched, left half), batched_a), Matmul(Slice(unbatched, right half), batched_b))
```
We also pull `Select` operations through elementwise operations, e.g.
```
Select(LinearCombination(a, b, ..), buckets)
-> LinearCombination(Select(a, buckets), Select(b, buckets), ..)
```

These changes result in either the same or marginally better performance, and we have traded the complex `SparseAffineDualActivate` operation for the much simpler `FusedPairwiseMulConcat`, and at the same time unlocked more flexibility (to e.g. feed difference weights and biases into the two perspectives).

Finally `TrainerBuilder` and associated `Trainer` is removed, after being deprecated for a while.